Debug and improve the kalman filter code
The Kalman gain was obtained through an integer division, so the result
was not valid. Also, the ROS publishing rate was corrected to a
reasonable value, and the whole publishing code was added to the kalman
class.

Two auxiliary scripts have been added for testing the Kalman filter.
They publish mock data of both the GPS and the marker.